Rename test file, move global to common, remove unused function
